# Conflicts:
#	src/main/java/mg/springboot/controller/AnnonceController.java
#	src/main/java/mg/springboot/repository/AnnonceRepository.java
#	src/main/java/mg/springboot/service/AnnonceService.java

+++ src/main/java/mg/springboot/controller/UtilisateurController.java
@@ -4,11 +4,17 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import lombok.Getter;
 import lombok.Setter;
+import mg.springboot.entity.Annonce;
 import mg.springboot.entity.Utilisateur;
+import mg.springboot.entity.Voiture;
+import mg.springboot.exception.AccessDeniedException;
+import mg.springboot.repository.VoitureRepository;
 import mg.springboot.security.Response;
 import mg.springboot.security.Token;
+import mg.springboot.service.AnnonceService;
 import mg.springboot.service.TokenService;
 import mg.springboot.service.UtilisateurService;
+import mg.springboot.service.VoitureService;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
@@ -20,11 +26,20 @@ import java.util.Optional;
 public class UtilisateurController {
     private final UtilisateurService utilisateurService;
     private final TokenService tokenService;
+    private final VoitureRepository voitureRepository;
+    private final VoitureService voitureService;
+    private final AnnonceService annonceService;
 
     public UtilisateurController(UtilisateurService utilisateurRepository,
-                                 TokenService tokenRepository) {
+                                 TokenService tokenRepository,
+                                 VoitureRepository voitureRepository,
+                                 VoitureService voitureService,
+                                 AnnonceService annonceService) {
         this.utilisateurService = utilisateurRepository;
         this.tokenService = tokenRepository;
+        this.voitureRepository = voitureRepository;
+        this.voitureService = voitureService;
+        this.annonceService = annonceService;
     }
 
     @PostMapping("/connexion")
@@ -39,8 +54,26 @@ public class UtilisateurController {
         return Response.send(HttpStatus.OK, "error", "Login ou mot de passe incorrect");
     }
 
-    @PostMapping("/inscription")
-    public Response<?> inscription(@Valid Utilisateur utilisateur) {
+    @PostMapping("/tokens/utilisateur")
+    public Response<?> connexionUtilisateur(HttpServletRequest request, String email, String motDePasse) {
+        Token token = tokenService.getToken(request);
+        if (token != null)
+            throw new AccessDeniedException("Vous êtes déjà connecté");
+        Optional<Utilisateur> utilisateur = utilisateurService.findByEmailAndMotDePasse(email, motDePasse);
+        if (utilisateur.isPresent()) {
+            Utilisateur utilisateurR = utilisateur.get();
+            if (utilisateurR.getLevel() != Utilisateur.LEVEL_USER)
+                return Response.send(HttpStatus.OK, "error", "Vous n'êtes pas utilisateur! Allez sur la page d'administration pour vous connecter");
+            return Response.send(HttpStatus.OK, "success","Connexion réussie", tokenService.createFor(utilisateurR));
+        }
+        return Response.send(HttpStatus.OK, "error", "Login ou mot de passe incorrect");
+    }
+
+    @PostMapping("/auth/utilisateurs")
+    public Response<?> inscription(HttpServletRequest httpServletRequest, @Valid Utilisateur utilisateur) {
+        Token token = tokenService.getToken(httpServletRequest);
+        if (token != null)
+            throw new AccessDeniedException("Vous êtes déjà connecté");
         if(utilisateur.getLevel() != Utilisateur.LEVEL_USER)
             return Response.send(HttpStatus.OK, "error", "Vous ne pouvez s'incrire que comme utilisateur");
         return Response.send(HttpStatus.OK, "success", "Inscription réussie", utilisateurService.save(utilisateur));
@@ -49,6 +82,15 @@ public class UtilisateurController {
     @DeleteMapping("/deconnexion")
     public Response<?> deconnexion(HttpServletRequest request) {
         Token token = tokenService.getToken(request);
+        tokenService.delete(token);
+        return Response.send(HttpStatus.OK, "success", "Déconnexion réussie", token);
+    }
+
+    @DeleteMapping("/tokens/utilisateur")
+    public Response<?> deconnexionUtilisateur(HttpServletRequest request) {
+        Token token = tokenService.getToken(request);
+        if (token == null)
+            throw new AccessDeniedException("Vous n'êtes pas connecté");
         tokenService.delete(token);
         return Response.send(HttpStatus.OK, "success", "Déconnexion réussie", token);
     }
@@ -97,5 +139,51 @@ public class UtilisateurController {
     public Response<?> modifyMotDePasse(@PathVariable String id, String motDePasse, String nouveauMotDePasse) {
         return Response.send(HttpStatus.OK, "success", "Le mot de passe a été modifié",
                     utilisateurService.modifyMotDePasse(id, motDePasse, nouveauMotDePasse));
+    }
+
+    @GetMapping("/api/voitures")
+    public Response<?> getVoitures(HttpServletRequest httpServletRequest) {
+        Token token = tokenService.getToken(httpServletRequest);
+        if (token == null)
+            throw new AccessDeniedException("Vous n'êtes pas connecté");
+        return Response.send(HttpStatus.OK, "success", utilisateurService.getVoitures(token.getUtilisateur().getId()));
+    }
+
+    @PostMapping("/api/voitures")
+    public Response<?> addVoiture(HttpServletRequest httpServletRequest, @Valid Voiture voiture) {
+        Token token = tokenService.getToken(httpServletRequest);
+        if (token.getUtilisateur().getId().equals(voiture.getUtilisateur().getId()))
+            return Response.send(HttpStatus.OK, "success", "La voiture a été ajoutée",
+                        voitureRepository.save(voiture));
+        throw new AccessDeniedException("Vous n'avez pas le droit d'ajouter cette voiture à cet utilisateur");
+    }
+
+    @PutMapping("/api/voitures/{id}")
+    public Response<?> modifyVoiture(HttpServletRequest httpServletRequest, @PathVariable Integer id, Voiture voiture) {
+        Token token = tokenService.getToken(httpServletRequest);
+        Voiture voitureR = voitureRepository.findById(id).get();
+        if (token.getUtilisateur().getId().equals(voitureR.getUtilisateur().getId()))
+            return Response.send(HttpStatus.OK, "success", "La voiture a été modifiée",
+                        voitureRepository.save(voiture));
+        throw new AccessDeniedException("Vous n'avez pas le droit de modifier cette voiture");
+    }
+
+    @DeleteMapping("/api/voitures/{id}")
+    public Response<?> deleteVoiture(HttpServletRequest httpServletRequest, @PathVariable Integer id) {
+        Token token = tokenService.getToken(httpServletRequest);
+        Voiture voitureR = voitureRepository.findById(id).get();
+        if (token.getUtilisateur().getId().equals(voitureR.getUtilisateur().getId()))
+            return Response.send(HttpStatus.OK, "success", "La voiture a été supprimée",
+                        voitureService.delete(id));
+        throw new AccessDeniedException("Vous n'avez pas le droit de supprimer cette voiture");
+    }
+
+    @PostMapping("/api/annonces")
+    public Response<?> addAnnonce(HttpServletRequest httpServletRequest, @Valid Annonce annonce) {
+        Token token = tokenService.getToken(httpServletRequest);
+        if (token.getUtilisateur().getId().equals(annonce.getVoiture().getUtilisateur().getId()))
+            return Response.send(HttpStatus.OK, "success", "L'annonce a été ajoutée",
+                        annonceService.save(annonce));
+        throw new AccessDeniedException("Vous n'avez pas le droit d'ajouter cette annonce à cet utilisateur");
     }
 }
